Created job to generate preview for every item or for a specific one.
